navigation_drawer.xml layout has been added
