+++ app/src/main/java/com/example/signupui/LoginActivity.java
@@ -1,15 +1,11 @@
 package com.example.signupui;
 
-import static android.content.ContentValues.TAG;
-
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
@@ -97,7 +93,7 @@ public class LoginActivity extends AppCompatActivity {
         finish();
     }
     void gotoHome(){
-        Intent intent =new Intent(getApplicationContext(), HomeActivity.class);
+        Intent intent =new Intent(getApplicationContext(), HomeActivitySample.class);
         startActivity(intent);
         finish();
     }
