+++ app/src/main/java/com/example/signupui/MainActivity.java
@@ -1,14 +1,10 @@
 package com.example.signupui;
 
-import static android.content.ContentValues.TAG;
-
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
@@ -108,7 +104,7 @@ public class MainActivity extends AppCompatActivity {
                 });
     }
      void gotoHome(){
-        Intent intent = new Intent(this, HomeActivity.class);
+        Intent intent = new Intent(this, HomeActivitySample.class);
         startActivity(intent);
         finish();
     }
